[SCredentialRequest] service added.
