+++ package.json
@@ -56,7 +56,7 @@
     "@mikro-orm/core": "^4.4.2",
     "@mikro-orm/postgresql": "^4.4.2",
     "@mikro-orm/reflection": "^4.4.2",
-    "@unumid/server-sdk": "git+ssh://git@github.com/UnumID/server-sdk-typescript.git#3.11.0",
+    "@unumid/server-sdk": "git+ssh://git@github.com/UnumID/server-sdk-typescript.git#meta",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
@@ -84,7 +84,7 @@
     "@typescript-eslint/eslint-plugin": "^4.15.0",
     "@typescript-eslint/parser": "^4.15.0",
     "@unumid/demo-types": "git+ssh://git@github.com/UnumID/demo-types#3.0.8",
-    "@unumid/types": "git+ssh://git@github.com/UnumID/types#3.24.1",
+    "@unumid/types": "git+ssh://git@github.com/UnumID/types#credStatus",
     "axios": "^0.21.1",
     "eslint": "^7.13.0",
     "eslint-config-semistandard": "^15.0.1",
