+++ src/services/api/credentialRequest/credentialRequest.hooks.ts
@@ -1,0 +1,51 @@
+import { BadRequest, GeneralError } from '@feathersjs/errors';
+import { Hook } from '@feathersjs/feathers';
+import { convertCredentialSubject, issueCredential as sdkIssueCredential, UnumDto, issueCredentials as sdkIssueCredentials } from '@unumid/server-sdk';
+import { CredentialData, CredentialPb, CredentialSubject, ProofPb } from '@unumid/types';
+import { Service as MikroOrmService } from 'feathers-mikro-orm';
+
+import { User } from '../../../entities/User';
+import { CredentialEntity, CredentialEntityOptions } from '../../../entities/Credential';
+import logger from '../../../logger';
+import { IssuerEntity } from '../../../entities/Issuer';
+import { CredentialStatus } from '@unumid/types/build/protos/credential';
+
+export const getDefaultIssuerEntity: Hook = async (ctx) => {
+  const issuerDataService = ctx.app.service('issuerData');
+  let defaultIssuerEntity: IssuerEntity;
+  try {
+    defaultIssuerEntity = await issuerDataService.getDefaultIssuerEntity();
+
+    return {
+      ...ctx,
+      params: {
+        ...ctx.params,
+        defaultIssuerEntity
+      }
+
+    };
+  } catch (e) {
+    logger.error('getDefaultIssuerEntity hook caught an error thrown by issuerDataService.getDefaultIssuerEntity', e);
+    throw e;
+  }
+};
+
+export const validateRequest: Hook = async (ctx) => {
+  const { params } = ctx;
+
+  if (!params.headers?.version) {
+    logger.info('CredentialRequest request made without version');
+  } else {
+    logger.info(`CredentialRequest request made with version ${params.headers?.version}`);
+  }
+
+  return ctx;
+};
+
+export const hooks = {
+  before: {
+    all: [validateRequest],
+    create: [getDefaultIssuerEntity]
+  },
+  after: {}
+};
