+++ src/services/index.ts
@@ -6,6 +6,7 @@ import pushTokenData from './data/pushToken.data.service';
 import user from './api/user/user.service';
 import issuer from './api/issuer/issuer.service';
 import credential from './api/credential/credential.service';
+import credentialRequest from './api/credentialRequest/credentialRequest.service';
 import pushToken from './api/pushToken/pushToken.service';
 import pushNotification from './api/pushNotification/pushNotification.service';
 
@@ -14,6 +15,7 @@ export default function (app: Application): void {
   app.configure(userData);
   app.configure(issuerData);
   app.configure(credentialData);
+  app.configure(credentialRequest);
   app.configure(pushTokenData);
   app.configure(user);
   app.configure(issuer);
